4lab/requirements.txt
astroid==2.9.0
asttokens==2.0.5
attrs==19.3.0
Automat==0.8.0
autopep8==1.6.0
backcall==0.2.0
black==22.3.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
cloud-init==21.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
cupy-cuda116==10.5.0
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.4
executing==0.8.3
fastrlock==0.8
fonttools==4.33.3
greenlet==1.1.2
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
idx2numpy==1.2.3
importlib-metadata==1.5.0
incremental==16.10.1
ipykernel==6.13.0
ipython==8.3.0
isort==5.10.1
jedi==0.18.1
Jinja2==2.10.1
joblib==1.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==7.3.1
jupyter-core==4.10.0
keyring==18.0.1
kiwisolver==1.4.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.7.1
llvmlite==0.38.1
macaroonbakery==1.3.1
MarkupSafe==1.1.0
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
more-itertools==4.2.0
mypy-extensions==0.4.3
nest-asyncio==1.5.5
netifaces==0.10.4
numba==0.55.2
numpy==1.22.3
oauthlib==3.1.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
prompt-toolkit==3.0.29
protobuf==3.6.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.8.0
pycups==1.9.73
Pygments==2.12.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.12.2
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.1
PyYAML==5.3.1
pyzmq==22.3.0
regex==2021.9.30
requests==2.22.0
requests-unixsocket==0.2.0
scikit-learn==1.1.1
scipy==1.8.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sklearn==0.0
sos==4.3
SQLAlchemy==1.4.29
ssh-import-id==5.10
stack-data==0.2.0
systemd-python==234
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.0
Twisted==18.9.0
typing-extensions==4.2.0
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.5
wrapt==1.13.3
xdot==1.2
zipp==1.0.0
zope.interface==4.7.1
